+++ spec/ig_markets/instrument_spec.rb
@@ -1,4 +1,8 @@
 describe IGMarkets::Instrument do
+  it 'accepts DFB as an expiry' do
+    expect(build(:instrument, expiry: 'DFB').expiry).to eq(nil)
+  end
+
   it 'adjusts API attributes correctly' do
     expect(IGMarkets::Instrument::OpeningHours.adjusted_api_attributes({})).to eq({})
     expect(IGMarkets::Instrument::OpeningHours.adjusted_api_attributes(market_times: [])).to eq([])